uploader-api/requirements.txt
babel==2.17.0 ; python_version >= "3.14" and python_version < "4.0"
backrefs==5.9 ; python_version >= "3.14" and python_version < "4.0"
bcrypt==5.0.0 ; python_version >= "3.14" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.14" and python_version < "4.0"
bump2version==1.0.1 ; python_version >= "3.14" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.14" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.14" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.14" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.14" and python_version < "4.0"
cloudpickle==3.1.2 ; python_version >= "3.14" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.14" and python_version < "4.0"
coverage==7.12.0 ; python_version >= "3.14" and python_version < "4.0"
elementpath==5.0.4 ; python_version >= "3.14" and python_version < "4.0"
flask-cors==6.0.1 ; python_version >= "3.14" and python_version < "4.0"
flask==3.1.2 ; python_version >= "3.14" and python_version < "4.0"
ghp-import==2.1.0 ; python_version >= "3.14" and python_version < "4.0"
griffe==1.15.0 ; python_version >= "3.14" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.14" and python_version < "4.0"
html5rdf==1.2.1 ; python_version >= "3.14" and python_version < "4"
idna==3.11 ; python_version >= "3.14" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.14" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.14" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.14" and python_version < "4.0"
lazydocs==0.4.8 ; python_version >= "3.14" and python_version < "4.0"
markdown-it-py==4.0.0 ; python_version >= "3.14" and python_version < "4.0"
markdown==3.10 ; python_version >= "3.14" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.14" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.14" and python_version < "4.0"
mergedeep==1.3.4 ; python_version >= "3.14" and python_version < "4.0"
mkdocs-autorefs==1.4.3 ; python_version >= "3.14" and python_version < "4.0"
mkdocs-get-deps==0.2.0 ; python_version >= "3.14" and python_version < "4.0"
mkdocs-material-extensions==1.3.1 ; python_version >= "3.14" and python_version < "4.0"
mkdocs-material==9.6.21 ; python_version >= "3.14" and python_version < "4.0"
mkdocs==1.6.1 ; python_version >= "3.14" and python_version < "4.0"
mkdocstrings-python==1.19.0 ; python_version >= "3.14" and python_version < "4.0"
mkdocstrings==0.30.1 ; python_version >= "3.14" and python_version < "4.0"
nanoid==2.0.0 ; python_version >= "3.14" and python_version < "4.0"
oldaplib==0.3.23 ; python_version >= "3.14" and python_version < "4.0"
owlrl==7.1.4 ; python_version >= "3.14" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.14" and python_version < "4.0"
paginate==0.5.7 ; python_version >= "3.14" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.14" and python_version < "4.0"
platformdirs==4.5.0 ; python_version >= "3.14" and python_version < "4.0"
prettytable==3.17.0 ; python_version >= "3.14" and python_version < "4"
pycparser==2.23 ; python_version >= "3.14" and python_version < "4.0" and implementation_name != "PyPy"
pygments==2.19.2 ; python_version >= "3.14" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.14" and python_version < "4.0"
pymdown-extensions==10.17.2 ; python_version >= "3.14" and python_version < "4.0"
pyparsing==3.2.5 ; python_version >= "3.14" and python_version < "4.0"
pyshacl==0.30.1 ; python_version >= "3.14" and python_version < "4"
pystrict==1.3 ; python_version >= "3.14" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.14" and python_version < "4.0"
pyvips==3.0.0 ; python_version >= "3.14" and python_version < "4.0"
pyyaml-env-tag==1.1 ; python_version >= "3.14" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.14" and python_version < "4.0"
rdflib==7.5.0 ; python_version >= "3.14" and python_version < "4.0"
rdflib[html]==7.5.0 ; python_version >= "3.14" and python_version < "4"
redis==6.4.0 ; python_version >= "3.14" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.14" and python_version < "4.0"
rich==14.2.0 ; python_version >= "3.14" and python_version < "4.0"
setuptools==80.9.0 ; python_version >= "3.14" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.14" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.14" and python_version < "4.0"
typer==0.20.0 ; python_version >= "3.14" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.14" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.14" and python_version < "4.0"
validators==0.35.0 ; python_version >= "3.14" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.14" and python_version < "4.0"
wcwidth==0.2.14 ; python_version >= "3.14" and python_version < "4"
werkzeug==3.1.4 ; python_version >= "3.14" and python_version < "4.0"
xmlschema==4.2.0 ; python_version >= "3.14" and python_version < "4.0"
yamale==6.1.0 ; python_version >= "3.14" and python_version < "4.0"
